Update 'Roles' migration (API)
- Change 'permission' column type to smallint

+++ api/database/migrations/2014_10_12_000001_create_roles_table.php
@@ -16,7 +16,7 @@ class CreateRolesTable extends Migration
             $table->increments('id');
             $table->string('name')
                   ->unique();
-            $table->integer('permission')
+            $table->smallInteger('permission')
                   ->unsigned();
             $table->dateTime('dateCreated');
             $table->dateTime('dateUpdated');
